Fix Dog class misxup
